+++ src/Resources/assets/js/contao-slick-bundle-theme.es6.js
@@ -1,2 +1,1 @@
-require('../js/contao-slick-bundle.es6');
 require('../scss/contao-slick-bundle-theme.scss');