+++ src/Resources/assets/js/contao-slick-bundle.es6.js
@@ -1,29 +1,44 @@
-(function($) {
-    window.SLICK_BUNDLE = {
-        init: function() {
-            $('.slick').each(function() {
+global.$ = global.jQuery = require('jquery');
 
-                var $this = $(this),
+import 'slick-carousel';
+require('../scss/contao-slick-bundle.scss');
+
+($ => {
+
+    $.fn.randomize = function(selector) {
+        (selector ? this.find(selector) : this).parent().each(function() {
+            $(this).children(selector).sort(function() {
+                return Math.random() - 0.5;
+            }).detach().appendTo(this);
+        });
+
+        return this;
+    };
+
+    let slickBundle = {
+        init : function(){
+            $('.slick').each(function(){
+                let $this = $(this),
                     container = $this.find('.slick-container'),
                     total = container.children().length;
 
                 // initialize slick, if more than one slide
                 if (total > 1) {
-                    var config = container.data('slickConfig');
-                    var initCallback = container.data('slickInitCallback');
-                    var afterInitCallback = container.data('slickAfterInitCallback');
+                    let config = container.data('slickConfig');
+                    let initCallback = container.data('slickInitCallback');
+                    let afterInitCallback = container.data('slickAfterInitCallback');
 
                     // don't init sliders if inside a hidden collapse -> done in shown.bs.collapse event
                     if ($this.closest('.collapse').length <= 0 || $this.closest('.collapse').hasClass('show'))
                     {
                         if(typeof initCallback !== 'undefined'){
-                            window.SLICK_BUNDLE.executeFunctionByName(initCallback, window, [container]);
+                            slickBundle.executeFunctionByName(initCallback, window, [container]);
                         }
 
-                        var slick = $this.data('slick', container.not('.slick-initialized').slick(config));
+                        let slick = $this.data('slick', container.not('.slick-initialized').slick(config));
 
                         if(typeof afterInitCallback !== 'undefined'){
-                            window.SLICK_BUNDLE.executeFunctionByName(afterInitCallback, window, [slick, container]);
+                            slickBundle.executeFunctionByName(afterInitCallback, window, [slick, container]);
                         }
                     }
 
@@ -34,9 +49,9 @@
             });
         },
         executeFunctionByName : function(functionName, context, args) {
-            var namespaces = functionName.split(".");
-            var func = namespaces.pop();
-            for(var i = 0; i < namespaces.length; i++) {
+            let namespaces = functionName.split(".");
+            let func = namespaces.pop();
+            for(let i = 0; i < namespaces.length; i++) {
                 context = context[namespaces[i]];
             }
 
@@ -49,27 +64,16 @@
         }
     };
 
-
     $(document).ready(function() {
-        SLICK_BUNDLE.init();
+        slickBundle.init();
     });
 
     $(document).on('shown.bs.modal', '.modal', function() {
-        SLICK_BUNDLE.init();
+        slickBundle.init();
     });
 
     $(document).on('shown.bs.collapse', '.collapse', function() {
-        SLICK_BUNDLE.init();
+        slickBundle.init();
     });
-
-    $.fn.randomize = function(selector) {
-        (selector ? this.find(selector) : this).parent().each(function() {
-            $(this).children(selector).sort(function() {
-                return Math.random() - 0.5;
-            }).detach().appendTo(this);
-        });
-
-        return this;
-    };
 
 })(jQuery);